Decouple editor.py
Move `MenuNav` into `navigation/`
Move `editor.py` to `application/`
