+++ src/navigation/menu_bar.py
@@ -1,32 +1,15 @@
 import datetime
-import os
 from asyncio import ensure_future
 
-from prompt_toolkit.application import Application
 from prompt_toolkit.application.current import get_app
-from prompt_toolkit.filters import Condition
 from prompt_toolkit.key_binding import KeyBindings
 from prompt_toolkit.key_binding.key_processor import KeyPressEvent
-from prompt_toolkit.layout.containers import (
-    ConditionalContainer,
-    Float,
-    HSplit,
-    VSplit,
-    Window,
-    WindowAlign,
-)
-from prompt_toolkit.layout.controls import FormattedTextControl
-from prompt_toolkit.layout.layout import Layout
+from prompt_toolkit.layout.containers import Float
 from prompt_toolkit.layout.menus import CompletionsMenu
-from prompt_toolkit.lexers import PygmentsLexer
 from prompt_toolkit.search import start_search
 from prompt_toolkit.shortcuts import set_title
-from prompt_toolkit.styles import Style
-from prompt_toolkit.widgets import MenuContainer, MenuItem, SearchToolbar, TextArea
-from pygments.lexers.markup import MarkdownLexer
+from prompt_toolkit.widgets import MenuContainer, MenuItem
 
-from application.state import ApplicationState
-from constants import NOTES_DIR
 from custom_types import MessageDialog, PopUpDialog, ScrollMenuDialog, TextInputDialog
 
 
@@ -200,8 +183,8 @@ class MenuNav:
 
     def _show_scroll(self, title: str, text: str) -> None:
         """Shows a MessageDialog with a certain title and text"""
-
-        async def coroutine(self: ThoughtBox) -> None:
+        # Type annotation is ThoughtBox but can't due to partial initialization
+        async def coroutine(self: object) -> None:
             dialog = ScrollMenuDialog(self, title, text)
             await self.show_dialog_as_float(dialog)
 
@@ -209,8 +192,8 @@ class MenuNav:
 
     def show_message(self, title: str, text: str) -> None:
         """Shows about message"""
-
-        async def coroutine(self: ThoughtBox) -> None:
+        # Type annotation is ThoughtBox but can't due to partial initialization
+        async def coroutine(self: object) -> None:
             dialog = MessageDialog(title, text)
             await self.show_dialog_as_float(dialog)
 
@@ -298,82 +281,3 @@ class MenuNav:
             self.do_undo()
 
         return bindings
-
-
-class ThoughtBox(MenuNav):
-    """Thought Box - The minimalist note-taking app"""
-
-    def __init__(self):
-        self.application_state = ApplicationState()
-        self.search_toolbar = SearchToolbar()
-        self.text_field = TextArea(
-            lexer=PygmentsLexer(MarkdownLexer),
-            scrollbar=True,
-            search_field=self.search_toolbar,
-        )
-        # style of menu can def play around here
-        self.style = Style.from_dict(
-            {
-                "status": "reverse",
-                "shadow": "bg:#440044",
-            }
-        )
-        self.body = HSplit(
-            [
-                self.text_field,
-                self.search_toolbar,
-                ConditionalContainer(
-                    content=VSplit(
-                        [
-                            # No longer need the bottom title bar
-                            # Window(
-                            #     FormattedTextControl(get_status_bar_left_text),
-                            #     style="class:status",
-                            #     align=WindowAlign.LEFT,
-                            # ),
-                            Window(
-                                FormattedTextControl(self.get_statusbar_middle_text),
-                                style="class:status",
-                            ),
-                            Window(
-                                FormattedTextControl(self.get_statusbar_right_text),
-                                style="class:status.right",
-                                width=9,
-                                align=WindowAlign.RIGHT,
-                            ),
-                        ],
-                        height=1,
-                    ),
-                    filter=Condition(lambda: self.application_state.show_status_bar),
-                ),
-            ]
-        )
-        # Initialize super class to get self.root_container
-        super().__init__()
-
-        self.layout = Layout(self.root_container, focused_element=self.text_field)
-
-        self.application = Application(
-            layout=self.layout,
-            enable_page_navigation_bindings=True,
-            style=self.style,
-            mouse_support=True,
-            full_screen=True,
-        )
-
-    def get_statusbar_middle_text(self) -> None:
-        """Gets status bar opens menu"""
-        return " Press Ctrl-K to open menu. "
-
-    def get_statusbar_right_text(self) -> None:
-        """Get status bar for the right text?"""
-        return " {}:{}  ".format(
-            self.text_field.document.cursor_position_row + 1,
-            self.text_field.document.cursor_position_col + 1,
-        )
-
-    def run(self) -> None:
-        """Run the application"""
-        # Create notes directory
-        os.makedirs(NOTES_DIR, exist_ok=True)
-        self.application.run()
